Address third round of PR #496 review feedback (tminka)
- VineTests: use GaussianCopula.ThetaToTau instead of inlining (2/pi)arcsin(rho).
- VineTests: ClaytonSample now calls ClaytonCopula.Sample instead of duplicating it.
- Move the test-only CopulaFamily enum and CopulaFactory out of the Runtime
  assembly into the test project (CopulaTests), and drop the stale
  '(Not yet implemented)' notes for Clayton and Gumbel.
